refactor nn-meter-cli
