+++ setup.py
@@ -35,7 +35,7 @@ setup(
             'kernel_detector/fusion_lib/*.json'],
     },
     entry_points={
-        'console_scripts': ['nn-meter=nn_meter.utils.nn_meter_cli:nn_meter_cli'],
+        'console_scripts': ['nn-meter=nn_meter.utils.nn_meter_cli.interface:nn_meter_cli'],
     },
     install_requires=[
         'numpy', 'tqdm', 'networkx', 'requests', 'protobuf', 'PyYAML', 'scikit_learn', 'packaging', 'jsonlines'
